tests/wpt/tests/css/css-animations/parsing/animation-duration-invalid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Animations: parsing animation-duration with invalid values</title>
<link rel="help" href="https://drafts.csswg.org/css-animations/#propdef-animation-duration">
<meta name="assert" content="animation-duration supports only the grammar '<time> #'.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_invalid_value("animation-duration", '-3s');
test_invalid_value("animation-duration", '0');
test_invalid_value("animation-duration", 'infinite');
test_invalid_value("animation-duration", '1s 2s');

test_invalid_value("animation-duration", 'initial, 1s');
test_invalid_value("animation-duration", '1s, initial');
</script>
</body>
</html>
